Fix maxidx+1 int overflow in scan_positionals (Greptile P1)
A literal index near INT_MAX (e.g. arg(2147483647)) overflowed
maxidx+1 computing PositionalInfo::count as int, wrapping negative --
describe_funcobj conflates a negative count with "can't be pinned
down statically" and falls back to rendering "(...)", so help() wrongly
reported a dynamic contract for a positional count that was actually
known, just huge. Widen PositionalInfo::count and scan_positionals'
maxidx/idx from int to long so the arithmetic can't overflow, and
update describe_funcobj's truncation-tail format from %d to %ld to
match. Regression test added (funchelp.comt test 13d).

+++ src/ComTerp/funcobjscan.h
@@ -72,7 +72,10 @@ public:
     // that's keyword-symbol classification only, with no notion of arg(n)
     // at all.
     struct PositionalInfo {
-        int count;         // -1 means "count could not be pinned down statically"
+        long count;        // -1 means "count could not be pinned down statically";
+                            // long (not int) so a literal index near INT_MAX
+                            // doesn't overflow computing count = maxidx + 1
+                            // (Greptile, PR #337)
         boolean uses_narg; // true if narg() appears anywhere in the body --
                             // treated as a signal the body is variadic
                             // (loops over an arg(n) run bounded by narg()),
